Extract the draft deck export menu into a component.
